kpwwniiexe/requirements.txt
aiohttp==3.8.5
aiolimiter==1.1.0
aiosignal==1.3.1
altair==5.1.1
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
docx==0.2.4
environs==9.5.0
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.99.1
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.3.3
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.35
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
htbuilder==0.6.2
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
ibm-cloud-sdk-core==3.17.2
ibm-cos-sdk==2.13.2
ibm-cos-sdk-core==2.13.2
ibm-cos-sdk-s3transfer==2.13.2
ibm-generative-ai==0.2.5
ibm-watson==7.0.1
ibm-watson-machine-learning==1.0.321
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
install==1.3.5
InstructorEmbedding==1.0.1
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.5
langchain==0.0.236
langsmith==0.0.10
lomond==0.3.3
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minio==7.2.0
mistune==3.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
notebook==7.0.3
notebook_shim==0.2.3
numexpr==2.8.5
numpy==1.25.2
onnxruntime==1.15.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
plotly==5.15.0
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.12
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pymilvus==2.3.3
Pympler==1.0.1
pyparsing==3.0.9
pypdf==3.12.2
PyPDF2==3.0.1
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rouge==1.0.1
rpds-py==0.10.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.11.2
selenium==4.15.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
st-annotated-text==4.0.1
stack-data==0.6.2
starlette==0.27.0
streamlit==1.24.1
streamlit-chat==0.1.1
streamlit-space==0.1.5
streamlit-toggle==0.1.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.1
trio==0.23.1
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
